Move the SF2 code into a symmetric location
Since I did it first it was still in engine. Now it
is in sample/sf2_support/sf2_import just like the
other formats
