+++ src/CMakeLists.txt
@@ -34,6 +34,7 @@ add_library(${PROJECT_NAME} STATIC
         sample/exs_support/exs_import.cpp
         sample/multisample_support/multisample_import.cpp
         sample/sf2_support/sf2_compound_file_impl.cpp
+        sample/sf2_support/sf2_import.cpp
         sample/sfz_support/sfz_parse.cpp
         sample/sfz_support/sfz_import.cpp
