+++ src/engine/engine.cpp
@@ -41,6 +41,7 @@
 #include "selection/selection_manager.h"
 #include "sample/sfz_support/sfz_import.h"
 #include "sample/exs_support/exs_import.h"
+#include "sample/sf2_support/sf2_import.h"
 #include "sample/multisample_support/multisample_import.h"
 #include "infrastructure/user_defaults.h"
 #include "infrastructure/md5support.h"
@@ -635,7 +636,7 @@ void Engine::loadCompoundElementIntoSelectedPartAndGroup(const sample::compound:
         {
             // TODO ok this refresh and restart is a bit unsatisfactory
             messageController->stopAudioThreadThenRunOnSerial([this, p](const auto &) {
-                loadSf2MultiSampleIntoSelectedPart(p.sampleAddress.path, p.sampleAddress.preset);
+                sf2_support::importSF2(p.sampleAddress.path, *this, p.sampleAddress.preset);
                 messageController->restartAudioThreadFromSerial();
                 serializationSendToClient(messaging::client::s2c_send_pgz_structure,
                                           getPartGroupZoneStructure(), *messageController);
@@ -698,7 +699,7 @@ void Engine::loadSampleIntoSelectedPartAndGroup(const fs::path &p, int16_t rootK
     {
         // TODO ok this refresh and restart is a bit unsatisfactory
         messageController->stopAudioThreadThenRunOnSerial([this, p](const auto &) {
-            loadSf2MultiSampleIntoSelectedPart(p);
+            sf2_support::importSF2(p, *this, -1);
             messageController->restartAudioThreadFromSerial();
             serializationSendToClient(messaging::client::s2c_send_pgz_structure,
                                       getPartGroupZoneStructure(), *messageController);
@@ -922,289 +923,6 @@ void Engine::sendEngineStatusToClient() const
     ec.runningEnvironment = runningEnvironment;
     messaging::client::serializationSendToClient(messaging::client::s2c_engine_status, ec,
                                                  *messageController);
-}
-void Engine::loadSf2MultiSampleIntoSelectedPart(const fs::path &p, int preset)
-{
-    assert(messageController->threadingChecker.isSerialThread());
-
-    SCLOG("Loading " << p.u8string() << " ps=" << preset);
-
-    auto cng = messaging::MessageController::ClientActivityNotificationGuard(
-        "Loading SF2", *(getMessageController()));
-
-    try
-    {
-        auto riff = std::make_unique<RIFF::File>(p.u8string());
-        auto sf = std::make_unique<sf2::File>(riff.get());
-        auto md5 = infrastructure::createMD5SumFromFile(p);
-
-        auto pt = getSelectionManager()->selectedPart;
-
-        auto &part = getPatch()->getPart(pt);
-        int firstGroup = -1;
-
-        auto spc = 0;
-        auto epc = sf->GetPresetCount();
-
-        if (preset >= 0)
-        {
-            spc = preset;
-            epc = preset + 1;
-        }
-
-        std::map<sf2::Instrument *, int> instToGroup;
-
-        for (int pc = spc; pc < epc; ++pc)
-        {
-            auto *preset = sf->GetPreset(pc);
-            auto pnm = std::string(preset->GetName());
-
-            for (int i = 0; i < preset->GetRegionCount(); ++i)
-            {
-                auto *presetRegion = preset->GetRegion(i);
-                sf2::Instrument *instr = presetRegion->pInstrument;
-
-                if (instToGroup.find(instr) == instToGroup.end())
-                {
-                    auto gnum = part->addGroup() - 1;
-                    auto &group = part->getGroup(gnum);
-                    group->name = std::string(instr->GetName()) + " / " + pnm;
-                    instToGroup[instr] = gnum;
-                }
-                auto grpnum = instToGroup[instr];
-                auto &grp = part->getGroup(grpnum);
-
-                if (instr->pGlobalRegion)
-                {
-                    // TODO: Global Region
-                }
-                for (int j = 0; j < instr->GetRegionCount(); ++j)
-                {
-                    auto region = instr->GetRegion(j);
-                    auto sfsamp = region->GetSample();
-                    if (sfsamp == nullptr)
-                        continue;
-
-                    messageController->updateClientActivityNotification("Loading sample " +
-                                                                        std::to_string(j));
-                    auto sid = sampleManager->loadSampleFromSF2(p, sf.get(), pc, i, j);
-                    if (!sid.has_value())
-                        continue;
-                    sampleManager->getSample(*sid)->md5Sum = md5;
-
-                    if (firstGroup < 0)
-                        firstGroup = grpnum;
-                    auto zn = std::make_unique<engine::Zone>(*sid);
-                    zn->engine = this;
-
-                    if (region->overridingRootKey >= 0)
-                    {
-                        zn->mapping.rootKey = region->overridingRootKey;
-                    }
-                    else
-                    {
-                        zn->mapping.rootKey += sfsamp->OriginalPitch - 60;
-                    }
-
-                    auto noneOr = [](auto a, auto b, auto c) {
-                        if (a != sf2::NONE)
-                            return a;
-                        if (b != sf2::NONE)
-                            return b;
-                        return c;
-                    };
-
-                    auto lk =
-                        noneOr(region->loKey, presetRegion->loKey, (int)sfsamp->OriginalPitch);
-                    auto hk =
-                        noneOr(region->hiKey, presetRegion->hiKey, (int)sfsamp->OriginalPitch);
-                    zn->mapping.keyboardRange = {lk, hk};
-
-                    auto lv = noneOr(region->minVel, presetRegion->minVel, 0);
-                    auto hv = noneOr(region->maxVel, presetRegion->maxVel, 127);
-                    zn->mapping.velocityRange = {lv, hv};
-
-                    // SF2 pitch correction is a *signed* char in cents.
-                    auto pcv = static_cast<int>(static_cast<int8_t>(sfsamp->PitchCorrection));
-                    zn->mapping.pitchOffset = 1.0 * pcv / 100.0;
-                    if (!zn->attachToSample(*sampleManager))
-                    {
-                        SCLOG("ERROR: Can't attach to sample");
-                        return;
-                    }
-                    auto &znSD = zn->variantData.variants[0];
-
-                    auto pn = region->GetPan(presetRegion);
-                    // pan in SF2 is -64 to 63 so hackety hack a bit
-                    if (pn < 0)
-                        zn->mapping.pan = pn / 64.f;
-                    else
-                        zn->mapping.pan = pn / 63.f;
-
-                    using namespace std;
-                    auto s = sfsamp;
-                    auto reg = region;
-
-                    zn->mapping.pitchOffset +=
-                        reg->GetCoarseTune(presetRegion) + reg->GetFineTune(presetRegion) * 0.01;
-
-                    auto s2a = scxt::modulation::secondsToNormalizedEnvTime;
-
-                    auto sus2l = [](double sus) {
-                        auto db = -sus / 10;
-                        return pow(10.0, db * 0.05);
-                    };
-
-                    zn->egStorage[0].dly = s2a(reg->GetEG1PreAttackDelay(presetRegion));
-                    zn->egStorage[0].a = s2a(reg->GetEG1Attack(presetRegion));
-                    zn->egStorage[0].h = s2a(reg->GetEG1Hold(presetRegion));
-                    zn->egStorage[0].d = s2a(reg->GetEG1Decay(presetRegion));
-                    zn->egStorage[0].s = sus2l(reg->GetEG1Sustain(presetRegion));
-                    zn->egStorage[0].r = s2a(reg->GetEG1Release(presetRegion));
-
-                    auto GetValue = [](const auto &val) {
-                        if (val == sf2::NONE)
-                            return ToString("NONE");
-                        return ToString(val);
-                    };
-
-                    zn->egStorage[1].dly = s2a(reg->GetEG2PreAttackDelay(presetRegion));
-                    zn->egStorage[1].a = s2a(reg->GetEG2Attack(presetRegion));
-                    zn->egStorage[1].h = s2a(reg->GetEG2Hold(presetRegion));
-                    zn->egStorage[1].d = s2a(reg->GetEG2Decay(presetRegion));
-                    zn->egStorage[1].s = sus2l(reg->GetEG2Sustain(presetRegion));
-                    zn->egStorage[1].r = s2a(reg->GetEG2Release(presetRegion));
-
-                    if (reg->HasLoop)
-                    {
-                        znSD.loopActive = true;
-                        znSD.startLoop = reg->LoopStart;
-                        znSD.endLoop = reg->LoopEnd;
-                    }
-                    else if (presetRegion->HasLoop)
-                    {
-                        znSD.loopActive = true;
-                        znSD.startLoop = presetRegion->LoopStart;
-                        znSD.endLoop = presetRegion->LoopEnd;
-                    }
-
-                    /*
-                     * The SF2 spec says anything above 20khz is a flat response
-                     * and that the filter cutoff is specified in cents (so really
-                     * 100 * midi key in 12TET). That's a midi key between 135 and 136,
-                     * so lets be safe a shave a touch
-                     */
-                    if (reg->initialFilterFc >= 0 && reg->initialFilterFc < 134 * 100)
-                    {
-                        // This only runs with audio thread stopped
-                        getMessageController()->threadingChecker.bypassThreadChecks = true;
-                        zn->setProcessorType(0, dsp::processor::ProcessorType::proct_CytomicSVF);
-                        zn->processorStorage[0].floatParams[0] =
-                            (reg->initialFilterFc / 100.0) - 69.0;
-                        zn->processorStorage[0].floatParams[2] =
-                            std::clamp((reg->initialFilterQ / 100.0), 0., 1.);
-                        getMessageController()->threadingChecker.bypassThreadChecks = false;
-                    }
-
-                    int currentModRow{0};
-                    // SCLOG("Unimplemented SF2 Features Follow: " )
-                    if (reg->modEnvToPitch != 0)
-                    {
-                        zn->routingTable.routes[currentModRow].source =
-                            voice::modulation::MatrixEndpoints::Sources::eg2A;
-                        zn->routingTable.routes[currentModRow].target =
-                            voice::modulation::MatrixEndpoints::MappingTarget::pitchOffsetA;
-                        zn->routingTable.routes[currentModRow].depth =
-                            1.0 * reg->modEnvToPitch / 19200;
-                        zn->onRoutingChanged();
-                        currentModRow++;
-                    }
-                    if (reg->modEnvToFilterFc != 0)
-                    {
-                        SCLOG("\tModulation Envelope Generator Cutoff="
-                              << GetValue(reg->modEnvToFilterFc) << "cents");
-                    }
-
-                    if (reg->vibLfoToPitch > 0)
-                    {
-                        // Use LFO1 as the vibrato LFO
-                        zn->modulatorStorage[0].modulatorShape =
-                            modulation::ModulatorStorage::LFO_SINE;
-                        zn->modulatorStorage[0].rate =
-                            (reg->freqVibLfo + 3637.63165623) * 0.01 / 12.0;
-
-                        if (reg->delayVibLfo > -11999)
-                        {
-                            SCLOG("Unimplemented: Curve Delay");
-                            // zn->modulatorStorage[0].curveLfoStorage.useenv = true;
-                            // zn->modulatorStorage[0].curveLfoStorage.delay = something;
-                            // zn->modulatorStorage[0].curveLfoStorage.attack = 0.0;
-                            // zn->modulatorStorage[0].curveLfoStorage.release = 1.0;
-                        }
-
-                        // TODO: Fix this with a constant
-                        zn->routingTable.routes[currentModRow].source = {'znlf', 'outp', 0};
-                        zn->routingTable.routes[currentModRow].target =
-                            voice::modulation::MatrixEndpoints::MappingTarget::pitchOffsetA;
-                        zn->routingTable.routes[currentModRow].depth =
-                            1.0 * reg->vibLfoToPitch / 19200;
-                        zn->onRoutingChanged();
-                        currentModRow++;
-                    }
-
-                    if (reg->modulators.size() > 0)
-                    {
-                        SCLOG("Unsupported Modulators present in : " << instr->GetName()
-                                                                     << " region " << i);
-                        SCLOG("\tCount =" << reg->modulators.size() << "");
-
-                        for (int mi = 0; mi < reg->modulators.size(); mi++)
-                        {
-                            auto m = reg->modulators[mi];
-                            SCLOG("\tsrc=" << m.ModSrcOper.Type << " target=" << m.ModDestOper
-                                           << " depth=" << m.ModAmount << " trans="
-                                           << m.ModTransOper << " srcamt=" << m.ModAmtSrcOper.Type);
-                            // PrintModulatorItem(&reg->modulators[i]);
-                        }
-                    }
-
-                    if (reg->modLfoToFilterFc != 0 || reg->modLfoToPitch != 0)
-                    {
-                        SCLOG("Unsupported: Modulation LFO: Delay="
-                              << ::sf2::ToSeconds(reg->delayModLfo)
-                              << "s, Frequency=" << ::sf2::ToHz(reg->freqModLfo)
-                              << "Hz, LFO to Volume=" << (reg->modLfoToVolume / 10) << "dB"
-                              << ", LFO to Filter Cutoff=" << reg->modLfoToFilterFc
-                              << ", LFO to Pitch=" << reg->modLfoToPitch);
-                    }
-
-                    if (reg->exclusiveClass)
-                        SCLOG("Unsupported: Exclusive Class : " << reg->exclusiveClass << " in "
-                                                                << instr->GetName() << " region "
-                                                                << i);
-                    ;
-
-                    grp->addZone(zn);
-                }
-            }
-        }
-        selectionManager->selectAction(
-            {pt, firstGroup, firstGroup >= 0 ? 0 : -1, true, true, true});
-    }
-    catch (RIFF::Exception e)
-    {
-        messageController->reportErrorToClient("SF2 Load Error", e.Message);
-        return;
-    }
-    catch (const SCXTError &e)
-    {
-        messageController->reportErrorToClient("SF2 Load Error", e.what());
-        return;
-    }
-    catch (...)
-    {
-        return;
-    }
 }
 
 void Engine::onSampleRateChanged()
