+++ src/engine/engine.h
@@ -463,8 +463,6 @@ struct Engine : MoveableOnly<Engine>, SampleRateSupport
     void pasteZone(const selection::SelectionManager::ZoneAddress &);
     std::string zoneClipboard;
 
-    void loadSf2MultiSampleIntoSelectedPart(const fs::path &, int preset = -1);
-
     /*
      * OnRegister generate and send all the metdata the client needs
      */
